requirements.txt
groq>=0.9.0
instructor>=1.3.0
langgraph>=0.2.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
polars>=0.20.0
pydantic>=2.7.0
python-dotenv>=1.0.0
requests>=2.32.0
